refactor(mobile): split account details
- extract detail screen modules
- harden repeated route params
