+++ apps/mobile/features/financial-accounts/lib/route-params.ts
@@ -2,14 +2,16 @@ import { requireFinancialAccountId } from "@/shared/types/assertions";
 import type { FinancialAccountId } from "@/shared/types/branded";
 
 export function parseFinancialAccountRouteParam(
-  value: string | undefined
+  value: string | string[] | undefined
 ): FinancialAccountId | null {
-  if (!value || value.trim().length === 0) {
+  const routeValue = Array.isArray(value) ? value[0] : value;
+
+  if (!routeValue || routeValue.trim().length === 0) {
     return null;
   }
 
   try {
-    return requireFinancialAccountId(value);
+    return requireFinancialAccountId(routeValue);
   } catch {
     return null;
   }
